+++ tests/conformance/tutorial/00-readme-member-import.scala
@@ -1,0 +1,5 @@
+// EXPECT: Red
+enum Colour:
+  case Red, Green
+import Colour.*
+@main def run(): Unit = println(Red)