+++ tests/conformance/tutorial/01-introduction-hello-app.scala
@@ -1,0 +1,3 @@
+// EXPECT: Hello, protoScala!
+object Hello extends App:
+  println("Hello, protoScala!")